bertApi/requirements.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.1
django-cors-headers==3.13.0
django-import-export==2.7.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
faiss-cpu==1.7.2
filelock==3.4.2
huggingface-hub==0.12.0
idna==3.3
joblib==1.1.0
MarkupPy==1.14
nltk==3.6.7
numpy==1.22.1
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
pandas==1.4.0
Pillow==9.0.0
pycodestyle==2.8.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
regex==2022.1.18
requests==2.27.1
sacremoses==0.0.47
scikit-learn==1.0.2
scipy==1.7.3
sentence-transformers==2.2.2
sentencepiece==0.1.96
six==1.16.0
sqlparse==0.4.2
tablib==3.2.0
threadpoolctl==3.0.0
tokenizers==0.11.4
toml==0.10.2
torch==1.11.0
torchvision==0.12.0
tqdm==4.62.3
transformers==4.26.0
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.26.8
xlrd==2.0.1
xlwt==1.3.0
